+++ postgres-single/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Version bump for new base image
 * Update to postgresql15
+* Add backup script
 
 ---